Various
* C++17 support

* rewrites some of the p2p code sans boost to allow g++-7 to compile it

* a great portion follows https://github.com/turtlecoin/turtlecoin/pull/535

* use `uint64_t` to replace `size_t`

+++ include/INode.h
@@ -8,6 +8,7 @@
 
 #include <cstdint>
 #include <functional>
+#include <memory>
 #include <system_error>
 #include <vector>
 
@@ -24,7 +25,7 @@ namespace CryptoNote {
 class INodeObserver {
 public:
   virtual ~INodeObserver() {}
-  virtual void peerCountUpdated(size_t count) {}
+  virtual void peerCountUpdated(uint64_t count) {}
   virtual void localBlockchainUpdated(uint32_t height) {}
   virtual void lastKnownBlockHeightUpdated(uint32_t height) {}
   virtual void poolChanged() {}
@@ -64,7 +65,7 @@ public:
   virtual void init(const Callback& callback) = 0;
   virtual bool shutdown() = 0;
 
-  virtual size_t getPeerCount() const = 0;
+  virtual uint64_t getPeerCount() const = 0;
   virtual uint32_t getLastLocalBlockHeight() const = 0;
   virtual uint32_t getLastKnownBlockHeight() const = 0;
   virtual uint32_t getLocalBlockCount() const = 0;
