+++ src/PaymentGateService/PaymentServiceConfiguration.h
@@ -42,7 +42,7 @@ struct Configuration {
   bool testnet;
   bool printAddresses;
 
-  size_t logLevel;
+  uint64_t logLevel;
 };
 
 } //namespace PaymentService
