+++ src/Platform/Linux/System/Dispatcher.cpp
@@ -55,7 +55,7 @@ private:
 
 static_assert(Dispatcher::SIZEOF_PTHREAD_MUTEX_T == sizeof(pthread_mutex_t), "invalid pthread mutex size");
 
-const size_t STACK_SIZE = 64 * 1024;
+const uint64_t STACK_SIZE = 64 * 1024;
 
 };
